Changed BrowserRouter to HashRouter
This works and will allow you to update the github pages name/url and have it always work.  The URL is a bit ugly though, but we could add a backlog task to look for an alternative.  The other thing I can do is make the ‘home’ location ‘/vapi’ instead of ‘/‘ and then setup a redirect from ‘/‘ to ‘/vtapi’ but its hacky.